chore: 모델 위치 변경 및 리뷰 수정
